requirements.txt
future
emoji
beautifulsoup4
requests
sqlalchemy
tswift
coffeehouse
spamwatch
python-telegram-bot==11.1.0
PyYAML==5.1.2
numpy
urbandict
pyowm
wikipedia
kbbi
feedparser
gtts
alphabet_detector
tabulate
parsel
youtube_dl
hurry.filesize
tldextract
psycopg2-binary
feedparser
tswift
pynewtonmath
#memes vvv
spongemock
zalgo-text
geopy
nltk
psutil
bleach
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
#translator
googletrans
jikanpy
speedtest-cli
